# internal/api/docs/swagger.yaml
definitions:
...
  entity.CreateBulkProductsRequest:
    properties:
      products:
        items:
          $ref: '#/definitions/entity.CreateProductRequestBulk'
        type: array
    type: object
...
  entity.CreateProductRequestBulk:
    properties:
      bill_format:
        type: string
      incoming_price:
        type: number
      name:
        type: string
      standard_price:
        type: number
      total_count:
        type: integer
    type: object
...
  entity.Error:
    properties:
      message:
        type: string
    type: object
...
  products.BulkCreateResponse:
    properties:
      message:
        type: string
      products:
        items:
          $ref: '#/definitions/products.Product'
        type: array
      success:
        type: boolean
    type: object
...
  products.Error:
    properties:
      error:
        type: string
    type: object
...
  products.Message:
    properties:
      message:
        type: string
    type: object
...
  products.Product:
    properties:
      bill_format:
        type: string
      branch_id:
        description: Added branch_id
        type: string
      category_id:
        type: string
      company_id:
        description: Company ID added
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      image_url:
        type: string
      incoming_price:
        description: Changed to double
        type: number
      name:
        type: string
      standard_price:
        description: Changed to double
        type: number
      total_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server
  title: API Gateway
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a list of products filtered by branch, category, and other
        optional parameters
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Category ID to filter products
        in: query
        name: category_id
        type: string
      - description: Product name to filter by
        in: query
        name: name
        type: string
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a list of products
      tags:
      - Products
    post:
      consumes:
      - multipart/form-data
      description: Create a new product with the provided details, including an optional
        file upload for the product image
      parameters:
      - description: Upload product image (optional)
        in: formData
        name: file
        type: file
      - description: ID of the product category
        in: formData
        name: category_id
        required: true
        type: string
      - description: Name of the product
        in: formData
        name: name
        required: true
        type: string
      - description: Billing format of the product
        in: formData
        name: bill_format
        required: true
        type: string
      - description: Incoming price of the product
        in: formData
        name: incoming_price
        required: true
        type: number
      - description: Standard price of the product
        in: formData
        name: standard_price
        required: true
        type: number
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Product successfully created
          schema:
            $ref: '#/definitions/products.Product'
        "400":
          description: Invalid input or bad request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a product
      tags:
      - Products
    put:
      consumes:
      - multipart/form-data
      description: Update the details of an existing product by ID, with optional
        media upload
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Upload product image (optional)
        in: formData
        name: file
        type: file
      - description: Name of the product
        in: formData
        name: name
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: string
      - description: Billing format
        in: formData
        name: bill_format
        type: string
      - description: Incoming price
        in: formData
        name: incoming_price
        required: true
        type: number
      - description: Standard price
        in: formData
        name: standard_price
        required: true
        type: number
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an existing product
      tags:
      - Products
  /products/bulk/{category_id}:
    post:
      consumes:
      - application/json
      description: Create multiple products in bulk with the provided details
      parameters:
      - description: Bulk product creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateBulkProductsRequest'
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Bulk products successfully created
          schema:
            $ref: '#/definitions/products.BulkCreateResponse'
        "400":
          description: Invalid input or bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create multiple products
      tags:
      - Products
...
  /products/excel-upload/{category_id}:
    post:
      consumes:
      - multipart/form-data
      description: Upload an Excel file containing product data, process it, and create
        products in bulk
      parameters:
      - description: Excel file containing products data
        in: formData
        name: file
        required: true
        type: file
      - description: Sheet name of file
        in: formData
        name: sheet_name
        required: true
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Error'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload an Excel file and create products
      tags:
      - Products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"